requirements.txt
fastapi
uvicorn
python-multipart
scikit-learn
tensorflow
numpy
pillow
joblib
